+++ src/components/TestDrive/Modal/EmployeeTestDriveModal.tsx
@@ -12,7 +12,8 @@ import useLocalize from '@hooks/useLocalize';
 import useOnboardingMessages from '@hooks/useOnboardingMessages';
 import useOnyx from '@hooks/useOnyx';
 import usePersonalPolicy from '@hooks/usePersonalPolicy';
-import {initMoneyRequest, setMoneyRequestAmount, setMoneyRequestCreated, setMoneyRequestDescription, setMoneyRequestMerchant, setMoneyRequestParticipants} from '@libs/actions/IOU';
+import {setMoneyRequestAmount, setMoneyRequestCreated, setMoneyRequestDescription, setMoneyRequestMerchant, setMoneyRequestParticipants} from '@libs/actions/IOU';
+import {initMoneyRequest} from '@libs/actions/IOU/MoneyRequest';
 import {setMoneyRequestReceipt} from '@libs/actions/IOU/Receipt';
 import {verifyTestDriveRecipient} from '@libs/actions/Onboarding';
 import setTestReceipt from '@libs/actions/setTestReceipt';
