+++ src/hooks/useReceiptScanDrop.tsx
@@ -7,7 +7,8 @@ import {hasOnlyPersonalPolicies as hasOnlyPersonalPoliciesUtil, isPaidGroupPolic
 import {generateReportID, getPolicyExpenseChat, isSelfDM} from '@libs/ReportUtils';
 import {shouldRestrictUserBillableActions} from '@libs/SubscriptionUtils';
 import type {ReceiptFile} from '@pages/iou/request/step/IOURequestStepScan/types';
-import {initMoneyRequest, setMoneyRequestParticipantsFromReport} from '@userActions/IOU';
+import {setMoneyRequestParticipantsFromReport} from '@userActions/IOU';
+import {initMoneyRequest} from '@userActions/IOU/MoneyRequest';
 import {setMoneyRequestReceipt} from '@userActions/IOU/Receipt';
 import {buildOptimisticTransactionAndCreateDraft} from '@userActions/TransactionEdit';
 import CONST from '@src/CONST';
