+++ src/libs/actions/IOU/index.ts
@@ -1,15 +1,10 @@
-import {format} from 'date-fns';
 import type {NullishDeep, OnyxCollection, OnyxEntry, OnyxKey} from 'react-native-onyx';
 import Onyx from 'react-native-onyx';
 import type {ValueOf} from 'type-fest';
 import type {UpdateMoneyRequestParams} from '@libs/API/parameters';
-import DateUtils from '@libs/DateUtils';
 import DistanceRequestUtils from '@libs/DistanceRequestUtils';
 import {getGPSRoutes, getGPSWaypoints} from '@libs/GPSDraftDetailsUtils';
-import {formatCurrentUserToAttendee} from '@libs/IOUUtils';
 import Log from '@libs/Log';
-import Navigation from '@libs/Navigation/Navigation';
-import {getCustomUnitID} from '@libs/PerDiemRequestUtils';
 import {getDistanceRateCustomUnit} from '@libs/PolicyUtils';
 import {
     getReportOrDraftReport,
@@ -18,43 +13,21 @@ import {
     isPolicyExpenseChat as isPolicyExpenseChatReportUtil,
     isSelfDM,
 } from '@libs/ReportUtils';
-import {startSpan} from '@libs/telemetry/activeSpans';
 import {getCategoryTaxDetails, getDistanceInMeters, isOdometerDistanceRequest as isOdometerDistanceRequestTransactionUtils} from '@libs/TransactionUtils';
-import {getRemoveDraftTransactionsByIDsData, removeDraftTransactionsByIDs} from '@userActions/TransactionEdit';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
-import ROUTES from '@src/ROUTES';
 import type * as OnyxTypes from '@src/types/onyx';
 import type {Accountant, Attendee, Participant} from '@src/types/onyx/IOU';
-import type {CurrentUserPersonalDetails} from '@src/types/onyx/PersonalDetails';
 import type {Unit} from '@src/types/onyx/Policy';
 import type RecentlyUsedTags from '@src/types/onyx/RecentlyUsedTags';
 import type {OnyxData} from '@src/types/onyx/Request';
-import type {Comment, Receipt} from '@src/types/onyx/Transaction';
+import type {Receipt} from '@src/types/onyx/Transaction';
 import {isEmptyObject} from '@src/types/utils/EmptyObject';
 import type BaseTransactionParams from './types/BaseTransactionParams';
 import type RequestMoneyParticipantParams from './types/RequestMoneyParticipantParams';
 import type {GPSPoint} from './types/TrackExpenseTransactionParams';
 
 type IOURequestType = ValueOf<typeof CONST.IOU.REQUEST_TYPE>;
-
-type InitMoneyRequestParams = {
-    reportID: string;
-    policy?: OnyxEntry<OnyxTypes.Policy>;
-    personalPolicy: Pick<OnyxTypes.Policy, 'id' | 'type' | 'autoReporting' | 'outputCurrency'> | undefined;
-    isFromGlobalCreate?: boolean;
-    isFromFloatingActionButton?: boolean;
-    currentIouRequestType?: IOURequestType | undefined;
-    newIouRequestType: IOURequestType | undefined;
-    report: OnyxEntry<OnyxTypes.Report>;
-    parentReport: OnyxEntry<OnyxTypes.Report>;
-    currentDate: string | undefined;
-    lastSelectedDistanceRates?: OnyxEntry<OnyxTypes.LastSelectedDistanceRates>;
-    currentUserPersonalDetails: CurrentUserPersonalDetails;
-    isTrackDistanceExpense?: boolean;
-    hasOnlyPersonalPolicies: boolean;
-    draftTransactionIDs?: string[];
-};
 
 type UpdateMoneyRequestData<TKey extends OnyxKey> = {
     params: UpdateMoneyRequestParams;
@@ -291,218 +264,6 @@ function getMoneyRequestPolicyTags({
         parentChatReport?.policyID ??
         allReports?.[`${ONYXKEYS.COLLECTION.REPORT}${participant.reportID}`]?.policyID;
     return getPolicyTagsData(iouReportPolicyID) ?? {};
-}
-
-/**
- * Initialize expense info
- * @param reportID to attach the transaction to
- * @param policy
- * @param isFromGlobalCreate
- * @param iouRequestType one of manual/scan/distance
- * @param report the report to attach the transaction to
- * @param parentReport the parent report to attach the transaction to
- */
-function initMoneyRequest({
-    reportID,
-    policy,
-    personalPolicy,
-    isFromGlobalCreate,
-    isTrackDistanceExpense = false,
-    isFromFloatingActionButton,
-    currentIouRequestType,
-    newIouRequestType,
-    report,
-    parentReport,
-    currentDate,
-    lastSelectedDistanceRates,
-    currentUserPersonalDetails,
-    hasOnlyPersonalPolicies,
-    draftTransactionIDs,
-}: InitMoneyRequestParams) {
-    // Generate a brand new transactionID
-    const newTransactionID = CONST.IOU.OPTIMISTIC_TRANSACTION_ID;
-    const currency = policy?.outputCurrency ?? personalPolicy?.outputCurrency ?? CONST.CURRENCY.USD;
-
-    const created = currentDate ?? format(new Date(), 'yyyy-MM-dd');
-
-    // We remove draft transactions created during multi scanning if there are some
-    removeDraftTransactionsByIDs(draftTransactionIDs, true);
-
-    // in case we have to re-init money request, but the IOU request type is the same with the old draft transaction,
-    // we should keep most of the existing data by using the ONYX MERGE operation
-    if (currentIouRequestType === newIouRequestType) {
-        // so, we just need to update the reportID, isFromGlobalCreate, created, currency
-        Onyx.merge(`${ONYXKEYS.COLLECTION.TRANSACTION_DRAFT}${newTransactionID}`, {
-            reportID,
-            isFromGlobalCreate,
-            isFromFloatingActionButton,
-            created,
-            currency,
-            transactionID: newTransactionID,
-        });
-        return;
-    }
-
-    const comment: Comment = {
-        attendees: formatCurrentUserToAttendee(currentUserPersonalDetails, reportID),
-    };
-    let requestCategory: string | null = null;
-
-    // Set up initial distance expense state
-    if (
-        newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE ||
-        newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE_MAP ||
-        newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE_MANUAL ||
-        newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE_ODOMETER ||
-        newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE_GPS
-    ) {
-        if (!isFromGlobalCreate) {
-            const isPolicyExpenseChat = isPolicyExpenseChatReportUtil(report) || isPolicyExpenseChatReportUtil(parentReport);
-            const customUnitRateID = DistanceRequestUtils.getCustomUnitRateID({reportID, isPolicyExpenseChat, isTrackDistanceExpense, policy, lastSelectedDistanceRates});
-            comment.customUnit = {customUnitRateID, name: CONST.CUSTOM_UNITS.NAME_DISTANCE};
-        } else if (hasOnlyPersonalPolicies) {
-            comment.customUnit = {customUnitRateID: CONST.CUSTOM_UNITS.FAKE_P2P_ID, name: CONST.CUSTOM_UNITS.NAME_DISTANCE};
-        }
-        if (comment.customUnit) {
-            comment.customUnit.quantity = null;
-        }
-        if (newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE_MANUAL || newIouRequestType === CONST.IOU.REQUEST_TYPE.DISTANCE_ODOMETER) {
-            comment.waypoints = undefined;
-        } else {
-            comment.waypoints = {
-                waypoint0: {keyForList: 'start_waypoint'},
-                waypoint1: {keyForList: 'stop_waypoint'},
-            };
-        }
-    }
-
-    if (newIouRequestType === CONST.IOU.REQUEST_TYPE.PER_DIEM) {
-        comment.customUnit = {
-            attributes: {
-                dates: {
-                    start: DateUtils.getStartOfToday(),
-                    end: DateUtils.getStartOfToday(),
-                },
-            },
-        };
-        if (!isFromGlobalCreate) {
-            const {customUnitID, category} = getCustomUnitID(report, parentReport, policy);
-            comment.customUnit = {...comment.customUnit, customUnitID};
-            requestCategory = category ?? null;
-        }
-    }
-
-    const defaultMerchant = newIouRequestType === CONST.IOU.REQUEST_TYPE.MANUAL ? CONST.TRANSACTION.DEFAULT_MERCHANT : CONST.TRANSACTION.PARTIAL_TRANSACTION_MERCHANT;
-
-    const newTransaction = {
-        amount: 0,
-        comment,
-        created,
-        currency,
-        category: requestCategory,
-        iouRequestType: newIouRequestType,
-        reportID,
-        transactionID: newTransactionID,
-        isFromGlobalCreate,
-        isFromFloatingActionButton,
-        merchant: defaultMerchant,
-    };
-
-    // Store the transaction in Onyx and mark it as not saved so it can be cleaned up later
-    // Use set() here so that there is no way that data will be leaked between objects when it gets reset
-    Onyx.set(`${ONYXKEYS.COLLECTION.TRANSACTION_DRAFT}${newTransactionID}`, newTransaction);
-
-    return newTransaction;
-}
-
-function createDraftTransaction(transaction: OnyxTypes.Transaction) {
-    if (!transaction) {
-        return;
-    }
-
-    const newTransaction = {
-        ...transaction,
-    };
-
-    Onyx.set(`${ONYXKEYS.COLLECTION.TRANSACTION_DRAFT}${transaction.transactionID}`, newTransaction);
-}
-
-function clearMoneyRequest(transactionID: string, draftTransactionIDs: string[] | undefined, skipConfirmation = false) {
-    const onyxData: Record<string, null | boolean> = {
-        ...getRemoveDraftTransactionsByIDsData(draftTransactionIDs),
-        [`${ONYXKEYS.COLLECTION.SKIP_CONFIRMATION}${transactionID}`]: skipConfirmation,
-    };
-    Onyx.multiSet(onyxData as Parameters<typeof Onyx.multiSet>[0]);
-}
-
-function startMoneyRequest(
-    iouType: ValueOf<typeof CONST.IOU.TYPE>,
-    reportID: string,
-    draftTransactionIDs: string[] | undefined,
-    requestType?: IOURequestType,
-    skipConfirmation = false,
-    backToReport?: string,
-    isFromFloatingActionButton?: boolean,
-) {
-    const sourceRoute = Navigation.getActiveRoute();
-    startSpan(CONST.TELEMETRY.SPAN_OPEN_CREATE_EXPENSE, {
-        name: '/money-request-create',
-        op: CONST.TELEMETRY.SPAN_OPEN_CREATE_EXPENSE,
-        attributes: {
-            [CONST.TELEMETRY.ATTRIBUTE_IOU_TYPE]: iouType,
-            [CONST.TELEMETRY.ATTRIBUTE_IOU_REQUEST_TYPE]: requestType ?? 'unknown',
-            [CONST.TELEMETRY.ATTRIBUTE_REPORT_ID]: reportID,
-            [CONST.TELEMETRY.ATTRIBUTE_ROUTE_FROM]: sourceRoute || 'unknown',
-        },
-    });
-    clearMoneyRequest(CONST.IOU.OPTIMISTIC_TRANSACTION_ID, draftTransactionIDs, skipConfirmation);
-    if (isFromFloatingActionButton) {
-        Onyx.set(`${ONYXKEYS.COLLECTION.TRANSACTION_DRAFT}${CONST.IOU.OPTIMISTIC_TRANSACTION_ID}`, {isFromFloatingActionButton});
-    }
-    switch (requestType) {
-        case CONST.IOU.REQUEST_TYPE.MANUAL:
-            Navigation.navigate(ROUTES.MONEY_REQUEST_CREATE_TAB_MANUAL.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        case CONST.IOU.REQUEST_TYPE.SCAN:
-            Navigation.navigate(ROUTES.MONEY_REQUEST_CREATE_TAB_SCAN.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        case CONST.IOU.REQUEST_TYPE.DISTANCE:
-            Navigation.navigate(ROUTES.MONEY_REQUEST_CREATE_TAB_DISTANCE.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        case CONST.IOU.REQUEST_TYPE.TIME:
-            Navigation.navigate(ROUTES.MONEY_REQUEST_CREATE_TAB_TIME.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        default:
-            Navigation.navigate(ROUTES.MONEY_REQUEST_CREATE.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-    }
-}
-
-function startDistanceRequest(
-    iouType: ValueOf<typeof CONST.IOU.TYPE>,
-    reportID: string,
-    draftTransactionIDs: string[] | undefined,
-    requestType?: IOURequestType,
-    skipConfirmation = false,
-    backToReport?: string,
-    isFromFloatingActionButton?: boolean,
-) {
-    clearMoneyRequest(CONST.IOU.OPTIMISTIC_TRANSACTION_ID, draftTransactionIDs, skipConfirmation);
-    if (isFromFloatingActionButton) {
-        Onyx.set(`${ONYXKEYS.COLLECTION.TRANSACTION_DRAFT}${CONST.IOU.OPTIMISTIC_TRANSACTION_ID}`, {isFromFloatingActionButton});
-    }
-    switch (requestType) {
-        case CONST.IOU.REQUEST_TYPE.DISTANCE_MAP:
-            Navigation.navigate(ROUTES.DISTANCE_REQUEST_CREATE_TAB_MAP.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        case CONST.IOU.REQUEST_TYPE.DISTANCE_MANUAL:
-            Navigation.navigate(ROUTES.DISTANCE_REQUEST_CREATE_TAB_MANUAL.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        case CONST.IOU.REQUEST_TYPE.DISTANCE_GPS:
-            Navigation.navigate(ROUTES.DISTANCE_REQUEST_CREATE_TAB_GPS.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-            return;
-        default:
-            Navigation.navigate(ROUTES.DISTANCE_REQUEST_CREATE.getRoute(CONST.IOU.ACTION.CREATE, iouType, CONST.IOU.OPTIMISTIC_TRANSACTION_ID, reportID, backToReport));
-    }
 }
 
 function setMoneyRequestReceiptState(transactionID: string, isDraft: boolean, shouldStopSmartscan = false) {
@@ -830,10 +591,7 @@ function setMultipleMoneyRequestParticipantsFromReport(transactionIDs: string[],
 }
 
 export {
-    clearMoneyRequest,
-    createDraftTransaction,
     getIOURequestPolicyID,
-    initMoneyRequest,
     resetDraftTransactionsCustomUnit,
     setCustomUnitRateID,
     setGPSTransactionDraftData,
@@ -861,10 +619,8 @@ export {
     setMoneyRequestTaxRate,
     setMoneyRequestTaxValue,
     setMoneyRequestTaxRateValues,
-    startMoneyRequest,
     updateLastLocationPermissionPrompt,
     setMoneyRequestReimbursable,
-    startDistanceRequest,
     getAllPersonalDetails,
     getAllTransactions,
     getAllTransactionViolations,
