+++ src/pages/Share/SubmitDetailsPage.tsx
@@ -22,15 +22,8 @@ import useReportIsArchived from '@hooks/useReportIsArchived';
 import useReportOrReportDraft from '@hooks/useReportOrReportDraft';
 import useThemeStyles from '@hooks/useThemeStyles';
 import type {GpsPoint} from '@libs/actions/IOU';
-import {
-    getIOURequestPolicyID,
-    getMoneyRequestParticipantsFromReport,
-    initMoneyRequest,
-    setMoneyRequestBillable,
-    setMoneyRequestParticipantsFromReport,
-    setMoneyRequestReimbursable,
-    updateLastLocationPermissionPrompt,
-} from '@libs/actions/IOU';
+import {getIOURequestPolicyID, getMoneyRequestParticipantsFromReport, setMoneyRequestBillable, setMoneyRequestParticipantsFromReport, setMoneyRequestReimbursable, updateLastLocationPermissionPrompt} from '@libs/actions/IOU';
+import {initMoneyRequest} from '@libs/actions/IOU/MoneyRequest';
 import {setMoneyRequestReceipt} from '@libs/actions/IOU/Receipt';
 import {requestMoney, trackExpense} from '@libs/actions/IOU/TrackExpense';
 import DateUtils from '@libs/DateUtils';
