+++ tests/actions/IOUTest.ts
@@ -7,19 +7,8 @@ import type {OnyxCollection, OnyxEntry, OnyxMergeCollectionInput} from 'react-na
 import type {SearchQueryJSON, SearchStatus} from '@components/Search/types';
 import useOnyx from '@hooks/useOnyx';
 import {clearAllRelatedReportActionErrors} from '@libs/actions/ClearReportActionErrors';
-import {
-    initMoneyRequest,
-    resetDraftTransactionsCustomUnit,
-    setMoneyRequestAmount,
-    setMoneyRequestBillable,
-    setMoneyRequestCategory,
-    setMoneyRequestCreated,
-    setMoneyRequestDateAttribute,
-    setMoneyRequestDescription,
-    setMoneyRequestDistanceRate,
-    setMoneyRequestMerchant,
-    setMoneyRequestTag,
-} from '@libs/actions/IOU';
+import {resetDraftTransactionsCustomUnit, setMoneyRequestAmount, setMoneyRequestBillable, setMoneyRequestCategory, setMoneyRequestCreated, setMoneyRequestDateAttribute, setMoneyRequestDescription, setMoneyRequestDistanceRate, setMoneyRequestMerchant, setMoneyRequestTag} from '@libs/actions/IOU';
+import {initMoneyRequest} from '@libs/actions/IOU/MoneyRequest';
 import {putOnHold} from '@libs/actions/IOU/Hold';
 import {calculateDiffAmount} from '@libs/actions/IOU/MoneyRequestBuilder';
 import {handleNavigateAfterExpenseCreate} from '@libs/actions/IOU/NavigationHelpers';
