+++ tests/ui/components/IOURequestStepConfirmationPageTest.tsx
@@ -13,7 +13,7 @@ import ONYXKEYS from '@src/ONYXKEYS';
 import type {Policy, TaxRatesWithDefault} from '@src/types/onyx';
 import type Transaction from '@src/types/onyx/Transaction';
 import type {WaypointCollection} from '@src/types/onyx/Transaction';
-import * as IOU from '../../../src/libs/actions/IOU';
+import * as MoneyRequest from '../../../src/libs/actions/IOU/MoneyRequest';
 import * as Split from '../../../src/libs/actions/IOU/Split';
 import * as TrackExpense from '../../../src/libs/actions/IOU/TrackExpense';
 import createRandomPolicy from '../../utils/collections/policies';
@@ -60,10 +60,10 @@ jest.mock('@assets/emojis', () => {
 jest.mock('@libs/EmojiTrie', () => ({
     buildEmojisTrie: jest.fn(),
 }));
-jest.mock('@libs/actions/IOU', () => {
-    const actualNav = jest.requireActual<typeof IOU>('@libs/actions/IOU');
+jest.mock('@libs/actions/IOU/MoneyRequest', () => {
+    const actual = jest.requireActual<typeof MoneyRequest>('@libs/actions/IOU/MoneyRequest');
     return {
-        ...actualNav,
+        ...actual,
         startMoneyRequest: jest.fn(),
     };
 });
@@ -326,7 +326,7 @@ describe('IOURequestStepConfirmationPageTest', () => {
         await waitForBatchedUpdatesWithAct();
 
         // Then startMoneyRequest should not be called from IOURequestConfirmationPage.
-        expect(IOU.startMoneyRequest).not.toHaveBeenCalled();
+        expect(MoneyRequest.startMoneyRequest).not.toHaveBeenCalled();
     });
 
     it('should create a split expense for a scanned receipt', async () => {
